scripts/requirements.txt
# Requirements for AI Sports Analytics Planning Scripts

# YAML parsing for story frontmatter
PyYAML>=6.0

# Optional: For enhanced JSON handling
orjson>=3.8.0

# Optional: For better CLI experience  
click>=8.0.0
rich>=13.0.0

# Optional: For validation
cerberus>=1.3.0
